requirements.txt
# Alephium Node Sniffer / Proxy
pyyaml>=6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiosqlite>=0.19.0
httpx>=0.25.0
python-multipart>=0.0.6
